chore: move getFileByPath to payload/uploads
