+++ packages/payload/src/exports/uploads.ts
@@ -1,0 +1,1 @@
+export { default as getFileByPath } from '../uploads/getFileByPath.js'
