+++ test/plugin-seo/seed/index.ts
@@ -2,7 +2,7 @@ import type { Payload } from 'payload'
 import type { PayloadRequest } from 'payload/types'
 
 import path from 'path'
-import { getFileByPath } from 'payload/utilities'
+import { getFileByPath } from 'payload/uploads'
 
 import { mediaSlug } from '../shared.js'
